refactor: move Round helper to new pkg/tools package
Extract the Round(time.Duration) helper from pkg/results/auxiliary.go
into a new dedicated package pkg/tools/tools.go.

Update merged_result.go and thread_results_test.go to import and use
tools.Round instead of the package-local Round. Delete auxiliary.go.

This is a pure code move with no behavior change, enabling future
packages (e.g. pkg/test_run) to use Round without importing pkg/results.
